+++ moimmoim_springBoot_project/src/main/java/moimmoimProject/mapper/ProfileMapper.java
@@ -1,0 +1,11 @@
+package moimmoimProject.mapper;
+
+import moimmoimProject.domain.ProfileDo;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+@Mapper
+public interface ProfileMapper {
+
+    ProfileDo findByUserIdNum(@Param("userIdNum") Long userIdNum);
+}
