+++ moimmoim_springBoot_project/src/main/java/moimmoimProject/mapper/UserMapper.java
@@ -1,0 +1,14 @@
+package moimmoimProject.mapper;
+
+import moimmoimProject.domain.UserDo;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+@Mapper
+public interface UserMapper {
+
+    UserDo getUserByCredentials(@Param("userId") String userId, @Param("userPassword") String userPassword);
+
+
+    UserDo findByUserIdNum(@Param("userIdNum") Long userIdNum);
+}
