--- moimmoim_springBoot_project/src/main/java/moimmoimProject/mapper/userMapper/ProfileMapper.java
@@ -1,10 +1,0 @@
-package moimmoimProject.mapper.userMapper;
-
-import moimmoimProject.domain.user.ProfileDo;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface ProfileMapper {
-
-    ProfileDo findByUserIdNum(Long userIdNum);
-}
