+++ moimmoim_springBoot_project/src/main/java/moimmoimProject/repository/ProfileRepository.java
@@ -1,6 +1,6 @@
 package moimmoimProject.repository;
 
-import moimmoimProject.domain.user.ProfileDo;
+import moimmoimProject.domain.ProfileDo;
 
 public interface ProfileRepository {
