+++ moimmoim_springBoot_project/src/main/java/moimmoimProject/service/UserService.java
@@ -1,0 +1,15 @@
+package moimmoimProject.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import moimmoimProject.mapper.UserMapper;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class UserService {
+    private final UserMapper userMapper;
+
+
+}
