--- moimmoim_springBoot_project/src/main/java/moimmoimProject/service/userService/UserService.java
@@ -1,15 +1,0 @@
-package moimmoimProject.service.userService;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import moimmoimProject.mapper.userMapper.UserMapper;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class UserService {
-    private final UserMapper userMapper;
-
-
-}
