+++ moimmoim_springBoot_project/src/main/resources/moimmoimProject/mapper/userMapper.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="moimmoimProject.mapper.UserMapper">
+    <resultMap id="userResultMap" type="moimmoimProject.mapper.domain.UserDo">
+        <id column="USERID_NUM" property="userIdNum" jdbcType="BIGINT" />
+        <result column="USER_ID" property="userId" jdbcType="VARCHAR" />
+        <result column="USER_EMAIL" property="userEmail" jdbcType="VARCHAR" />
+        <result column="PASSWORD" property="userPassword" jdbcType="VARCHAR" />
+        <result column="USER_NAME" property="userName" jdbcType="VARCHAR" />
+        <result column="USER_NICKNAME" property="userNickname" jdbcType="VARCHAR" />
+        <result column="PHONE_NUM" property="phoneNum" jdbcType="VARCHAR" />
+        <result column="BIRTH" property="birth" jdbcType="VARCHAR" />
+        <result column="CATEGORY_NUM" property="categoryNum" jdbcType="INTEGER" />
+        <result column="HOST_LEVEL_NUM" property="hostLevelNum" jdbcType="INTEGER" />
+        <result column="LOCATION_NUM" property="location" jdbcType="VARCHAR" />
+        <result column="CREATE_DATE" property="createDate" jdbcType="DATE" />
+    </resultMap>
+
+    <select id="getUserByCredentials" parameterType="map" resultMap="userResultMap">
+        SELECT *
+        FROM USERS
+        WHERE USER_ID = #{userId}
+        AND PASSWORD = #{userPassword}
+    </select>
+
+    <select id="findByUserIdNum" parameterType="map" resultMap="userResultMap">
+        SELECT *
+        FROM USERS
+        WHERE USERID_NUM = #{userIdNum}
+    </select>
+</mapper>
