--- moimmoim_springBoot_project/src/main/resources/moimmoimProject/mapper/userMapper/userMapper.xml
@@ -1,12 +1,0 @@
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="moimmoimProject.mapper.userMapper.UserMapper">
-
-<!--    <select id="findByName" resultType="User">-->
-<!--        select * from users-->
-<!--    </select>-->
-
-
-</mapper>
